+Category adapter db prepare and add dialog category

+++ app/src/main/java/com/wirrez/shoppingcart/Item.java
@@ -6,13 +6,13 @@ package com.wirrez.shoppingcart;
  */
 
 public class Item {
-    int _id;
+    long _id;
     String _name;
     int _count;
     String _icon;
     Boolean _cross;
 
-    public Item(int id,String name,int count,String icon,Boolean cross)
+    public Item(long id,String name,int count,String icon,Boolean cross)
     {
         this._id = id;
         this._name = name;
